+++ next.config.js
@@ -9,6 +9,17 @@ module.exports = withCSS({
         use: "eslint-loader"
       });
     }
+
+    const originalEntry = config.entry;
+    config.entry = async () => {
+      const entries = await originalEntry();
+
+      if (entries["main.js"]) {
+        entries["main.js"].unshift("./client/polyfills.js");
+      }
+
+      return entries;
+    };
     return config;
   }
 });