+++ src/components/SummaryTable.js
@@ -434,7 +434,7 @@ const FoodActivitiesTable = props => (
   <React.Fragment>
     <Table.Row TITLE>
       <Table.CellHeader>
-        <Paragraph mb={0}>**Food activities**</Paragraph>
+        <Paragraph mb={0}>**Activities**</Paragraph>
       </Table.CellHeader>
       <Table.Cell>
         <div />
@@ -455,8 +455,34 @@ const FoodActivitiesTable = props => (
           </div>
         </Table.Cell>
         {props.hideChangeButtons ? null : (
-          <Table.Cell>
-            <div />
+          <Table.Cell className="summaryTableChangeCell">
+            <AnchorTag
+              id="changeActivitiesCustomersRow"
+              href="/edit/customer-type"
+            >
+              Change
+            </AnchorTag>
+          </Table.Cell>
+        )}
+      </Table.Row>
+    ) : null}
+
+    {props.import_export_activities ? (
+      <Table.Row id="activitiesBusinessImportExportRow">
+        <Table.CellHeader>Food activities</Table.CellHeader>
+        <Table.Cell className="summaryTableDataCell">
+          <div className="bold" id="import_export_activities">
+            {props.import_export_activities}
+          </div>
+        </Table.Cell>
+        {props.hideChangeButtons ? null : (
+          <Table.Cell className="summaryTableChangeCell">
+            <AnchorTag
+              id="changeActivitiesBusinessImportExportRow"
+              href="/edit/business-import-export"
+            >
+              Change
+            </AnchorTag>
           </Table.Cell>
         )}
       </Table.Row>
@@ -522,5 +548,6 @@ SummaryTable.propTypes = {
   establishment_street: PropTypes.string,
   establishment_town: PropTypes.string,
   establishment_postcode: PropTypes.string,
-  customer_type: PropTypes.string
+  customer_type: PropTypes.string,
+  import_export_activities: PropTypes.string
 };