+++ tests/contract/registration.contract.test.js
@@ -1,6 +1,6 @@
 jest.mock("../../src/server/config.js", () => ({
   SUBMIT_URL:
-    "https://register-a-food-business-service-dev-double.azurewebsites.net/api/registration/createNewRegistration"
+    "https://dev-register-a-food-business-service.azurewebsites.net/api/registration/createNewRegistration"
 }));
 const {
   sendRequest
@@ -12,7 +12,7 @@ const validBody = {
         establishment_trading_name: "Itsu",
         establishment_primary_number: "329857245",
         establishment_secondary_number: "84345245",
-        establishment_email: "django@uk.ibm.com",
+        establishment_email: "fsatestemail.valid@gmail.com",
         establishment_opening_date: "2018-06-07"
       },
       operator: {
@@ -23,7 +23,7 @@ const validBody = {
         operator_street: "Some St.",
         operator_town: "London",
         operator_primary_number: "9827235",
-        operator_email: "operator@email.com",
+        operator_email: "fsatestemail.valid@gmail.com",
         operator_type: "Sole trader"
       },
       premise: {
@@ -36,7 +36,8 @@ const validBody = {
       activities: {
         customer_type: "End consumer",
         business_type: "Livestock farm",
-        business_type_search_term: "Example"
+        business_type_search_term: "Example",
+        import_export_activities: "None"
       }
     },
     metadata: {
@@ -78,7 +79,8 @@ const invalidBody = {
       activities: {
         customer_type: "End consumer",
         business_type: "Livestock farm",
-        business_type_search_term: "Example"
+        business_type_search_term: "Example",
+        import_export_activities: "None"
       }
     },
     metadata: {
@@ -110,12 +112,17 @@ describe("Registration contract", () => {
       expect(typeof realJsonResponse["fsa-rn"]).toBe(
         typeof doubleJsonResponse["fsa-rn"]
       );
-      expect(realJsonResponse.tascomiResponse).toEqual(
-        doubleJsonResponse.tascomiResponse
+      expect(typeof realJsonResponse.tascomiResponse.id).toEqual(
+        typeof doubleJsonResponse.tascomiResponse.id
+      );
+      expect(typeof realJsonResponse.tascomiResponse.online_reference).toEqual(
+        typeof doubleJsonResponse.tascomiResponse.online_reference
       );
       expect(realJsonResponse.reg_submission_date).toEqual(
         doubleJsonResponse.reg_submission_date
       );
+      expect(realJsonResponse.email_fbo).toEqual(doubleJsonResponse.email_fbo);
+      expect(realJsonResponse.email_lc).toEqual(doubleJsonResponse.email_lc);
     });
   });